Adding Epoch Block Table.
