+++ examples/explorer/src/components/Blocks/BlockTable/BlockTable.jsx
@@ -1,125 +1,64 @@
-import React, { useState } from 'react';
+import React from 'react';
 import Table from 'react-bootstrap/Table';
-import Pagination from 'react-bootstrap/Pagination';
 
 import graphql from 'babel-plugin-relay/macro';
-import { createRefetchContainer } from 'react-relay';
+import { createFragmentContainer } from 'react-relay';
 
-import BlockTablePage from '../BlockTablePage/BlockTablePage';
-import { TABLE_PAGE_SIZE } from '../../../helpers/constants';
+import BlockLink from '../../Commons/BlockLink/BlockLink';
+import EpochLink from '../../Commons/EpochLink/EpochLink';
 
-const getBlocksFromConnection = data => {
-  return data.allBlocks.edges.map(edge => edge.node);
+const orderBlocks = blocks => {
+  return blocks.sort((b1, b2) => Number(b2.chainLength) - Number(b1.chainLength));
 };
 
-const BlockTable = ({ data, relay }) => {
-  const [start, setStart] = useState(1);
-  const blocks = getBlocksFromConnection(data);
-  const { pageInfo } = data.allBlocks;
-
-  const handlePageChange = (vars, callback) => {
-    relay.refetch(
-      {
-        first: vars.first || null,
-        last: vars.last || null,
-        after: vars.after || null,
-        before: vars.before || null
-      },
-      error => {
-        if (error) {
-          console.error(error); // eslint-disable-line no-console
-        }
-        callback();
-      }
-    );
-  };
-
-  const openPreviousPage = () => {
-    handlePageChange({ before: pageInfo.startCursor, last: TABLE_PAGE_SIZE }, () =>
-      setStart(start - TABLE_PAGE_SIZE)
-    );
-  };
-
-  const openNextPage = () => {
-    handlePageChange({ after: pageInfo.endCursor, first: TABLE_PAGE_SIZE }, () =>
-      setStart(start + TABLE_PAGE_SIZE)
-    );
-  };
-
-  const openLastPage = () => {
-    handlePageChange({ last: TABLE_PAGE_SIZE }, () => setStart(start - TABLE_PAGE_SIZE));
-  };
-
-  const openFirstPage = () => {
-    handlePageChange({ first: TABLE_PAGE_SIZE }, () => setStart(0));
-  };
-
+const BlockTable = ({ blocks }) => {
+  const orderedBlocks = orderBlocks(blocks);
   return (
-    <>
-      <Table striped bordered hover>
-        <thead>
+    <Table striped bordered hover>
+      <thead>
+        <tr>
+          <th>Chain Length</th>
+          <th>Hash</th>
+          <th>Epoch</th>
+          <th>Slot</th>
+          <th>Tx count</th>
+        </tr>
+      </thead>
+      <tbody>
+        {orderedBlocks.map(block => (
           <tr>
-            <th>Chain Length</th>
-            <th>Hash</th>
-            <th>Epoch</th>
-            <th>Slot</th>
-            <th>Tx count</th>
+            <td>
+              <BlockLink chainLength={block.chainLength} />
+            </td>
+            <td>
+              <BlockLink id={block.id} />
+            </td>
+            <td>
+              <EpochLink number={block.date.epoch.id} />
+            </td>
+            <td>{block.date.slot}</td>
+            <td>{block.transactions.length}</td>
           </tr>
-        </thead>
-        <BlockTablePage {...{ blocks }} />
-      </Table>
-      <Pagination>
-        <Pagination.Item onClick={openLastPage} disabled={!pageInfo.hasNextPage}>
-          Last
-        </Pagination.Item>
-        <Pagination.Prev onClick={openNextPage} disabled={!pageInfo.hasNextPage} />
-        <Pagination.Ellipsis disabled />
-        <Pagination.Next onClick={openPreviousPage} disabled={!pageInfo.hasPreviousPage} />
-        <Pagination.Item onClick={openFirstPage} disabled={!pageInfo.hasPreviousPage}>
-          First
-        </Pagination.Item>
-      </Pagination>
-    </>
+        ))}
+      </tbody>
+    </Table>
   );
 };
 
-export default createRefetchContainer(
-  BlockTable,
-  {
-    data: graphql`
-      fragment BlockTable_data on Query
-        @argumentDefinitions(
-          first: { type: "Int" }
-          last: { type: "Int" }
-          after: { type: "BlockCursor" }
-          before: { type: "BlockCursor" }
-        ) {
-        allBlocks(first: $first, last: $last, after: $after, before: $before) {
-          edges {
-            cursor
-            node {
-              ...BlockTablePage_blocks
-            }
-          }
-          pageInfo {
-            hasNextPage
-            hasPreviousPage
-            endCursor
-            startCursor
-          }
-          totalCount
+export default createFragmentContainer(BlockTable, {
+  blocks: graphql`
+    fragment BlockTable_blocks on Block @relay(plural: true) {
+      id
+      date {
+        epoch {
+          id
         }
+        slot
       }
-    `
-  },
-  graphql`
-    query BlockTableRefetchQuery(
-      $first: Int
-      $last: Int
-      $after: BlockCursor
-      $before: BlockCursor
-    ) {
-      ...BlockTable_data @arguments(first: $first, last: $last, after: $after, before: $before)
+      chainLength
+      transactions {
+        id
+      }
     }
   `
-);
+});
