+++ examples/explorer/src/components/Epochs/EpochInfo/epochInfo.scss
@@ -2,11 +2,7 @@
 
 .epochInfo {
   display: flex;
-  overflow: hidden;
   flex-direction: column;
-  justify-content: center;
   align-items: center;
-  box-shadow: 0px 0px 6px 1px grey;
-  padding: 1%;
-  width: 80%;
+  width: 100%;
 }
